+++ packages/cli/test/dev/utils.ts
@@ -1,32 +1,37 @@
-const fs = require('fs-extra');
-const { join, resolve } = require('path');
-const _execa = require('execa');
-const fetch = require('node-fetch');
-const retry = require('async-retry');
-const { satisfies } = require('semver');
-const stripAnsi = require('strip-ansi');
-const {
-  fetchCachedToken,
-} = require('../../../../test/lib/deployment/now-deploy');
-const { spawnSync, execFileSync } = require('child_process');
+import fs from 'fs-extra';
+import { join, resolve } from 'path';
+import _execa, { ExecaChildProcess, type Options } from 'execa';
+import fetch, { type RequestInit, type Response } from 'node-fetch';
+import retry from 'async-retry';
+import { satisfies } from 'semver';
+import stripAnsi from 'strip-ansi';
+import { fetchCachedToken } from '../../../../test/lib/deployment/now-deploy';
+import { spawnSync, execFileSync } from 'child_process';
 
 jest.setTimeout(10 * 60 * 1000);
 
 const isCI = !!process.env.CI;
-const sleep = ms => new Promise(resolve => setTimeout(resolve, ms));
+
+export function sleep(ms: number) {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
 
 let port = 3000;
 
 const binaryPath = resolve(__dirname, `../../scripts/start.js`);
-const fixture = name => join('test', 'dev', 'fixtures', name);
-const fixtureAbsolute = name => join(__dirname, 'fixtures', name);
+
+export function fixture(name: string) {
+  return join('test', 'dev', 'fixtures', name);
+}
+
+const fixtureAbsolute = (name: string) => join(__dirname, 'fixtures', name);
 
 let processCounter = 0;
 const processList = new Map();
 
-function execa(...args) {
+function execa(initial: string, args: string[], options: Options<null> = {}) {
   const procId = ++processCounter;
-  const child = _execa(...args);
+  const child = _execa(initial, args, options);
 
   processList.set(procId, child);
   child.on('close', () => processList.delete(procId));
@@ -34,7 +39,12 @@ function execa(...args) {
   return child;
 }
 
-function fetchWithRetry(url, opts = {}) {
+type FetchOptions = RequestInit & {
+  status?: number;
+  retries?: number;
+};
+
+export function fetchWithRetry(url: string, opts: FetchOptions = {}) {
   return retry(
     async () => {
       const res = await fetch(url, opts);
@@ -57,23 +67,39 @@ function fetchWithRetry(url, opts = {}) {
   );
 }
 
-function createResolver() {
-  let resolver;
-  let rejector;
+type ResolverPromise<T> = Promise<T> & {
+  resolve: (value: PromiseLike<null> | null) => void;
+  reject: (reason?: any) => void;
+};
+
+function createResolver(): ResolverPromise<null> {
+  let resolver: ResolverPromise<null>['resolve'];
+  let rejector: ResolverPromise<null>['reject'];
+
   const p = new Promise((resolve, reject) => {
     resolver = resolve;
     rejector = reject;
-  });
+  }) as ResolverPromise<null>;
+
+  //@ts-expect-error
   p.resolve = resolver;
+  //@ts-expect-error
   p.reject = rejector;
+
   return p;
 }
 
-function formatOutput({ stderr, stdout }) {
+export function formatOutput({
+  stderr,
+  stdout,
+}: {
+  stderr: string;
+  stdout: string;
+}) {
   return `Received:\n"${stderr}"\n"${stdout}"`;
 }
 
-function printOutput(fixture, stdout, stderr) {
+function printOutput(fixture: string, stdout: string, stderr: string) {
   const lines = (
     `\nOutput for "${fixture}"\n` +
     `\n----- stdout -----\n` +
@@ -82,7 +108,7 @@ function printOutput(fixture, stdout, stderr) {
     stderr
   ).split('\n');
 
-  const getPrefix = nr => {
+  const getPrefix = (nr: number) => {
     return nr === 0 ? '╭' : nr === lines.length - 1 ? '╰' : '│';
   };
 
@@ -92,7 +118,7 @@ function printOutput(fixture, stdout, stderr) {
   );
 }
 
-function shouldSkip(name, versions) {
+export function shouldSkip(name: string, versions: string) {
   if (!satisfies(process.version, versions)) {
     // eslint-disable-next-line no-console
     console.log(`Skipping "${name}" because it requires "${versions}".`);
@@ -102,20 +128,22 @@ function shouldSkip(name, versions) {
   return false;
 }
 
-function validateResponseHeaders(res, podId) {
+export function validateResponseHeaders(res: Response, podId?: string) {
   if (res.status < 500) {
+    const cacheControlCount = res.headers.get('cache-control')?.length || 0;
+    expect(cacheControlCount > 0).toBeTruthy();
+
     expect(res.headers.get('server')).toEqual('Vercel');
-    expect(res.headers.get('cache-control').length > 0).toBeTruthy();
     expect(res.headers.get('x-vercel-id')).toBeTruthy();
+
     if (podId) {
-      expect(
-        res.headers.get('x-vercel-id').includes(`::${podId}-`)
-      ).toBeTruthy();
+      const vercelID = res.headers.get('x-vercel-id') || '';
+      expect(vercelID.includes(`::${podId}-`)).toBeTruthy();
     }
   }
 }
 
-async function exec(directory, args = []) {
+export async function exec(directory: string, args: string[] = []) {
   const token = await fetchCachedToken();
 
   // eslint-disable-next-line no-console
@@ -124,27 +152,19 @@ async function exec(directory, args = []) {
       process.env.VERCEL_TEAM_ID ? ' --scope ***' : ''
     } ${args.join(' ')}`
   );
-  return execa(
-    binaryPath,
-    [
-      'dev',
-      directory,
-      '-t',
-      token,
-      ...(process.env.VERCEL_TEAM_ID
-        ? ['--scope', process.env.VERCEL_TEAM_ID]
-        : []),
-      ...args,
-    ],
-    {
-      reject: false,
-      shell: true,
-      env: { __VERCEL_SKIP_DEV_CMD: 1 },
-    }
-  );
+
+  const scope: string[] = process.env.VERCEL_TEAM_ID
+    ? ['--scope', process.env.VERCEL_TEAM_ID]
+    : [];
+
+  return execa(binaryPath, ['dev', directory, '-t', token, ...scope, ...args], {
+    reject: false,
+    shell: true,
+    env: { __VERCEL_SKIP_DEV_CMD: '1' },
+  });
 }
 
-async function runNpmInstall(fixturePath) {
+async function runNpmInstall(fixturePath: string) {
   if (await fs.pathExists(join(fixturePath, 'package.json'))) {
     let command;
     if (await fs.pathExists(join(fixturePath, 'package-lock.json'))) {
@@ -160,18 +180,20 @@ async function runNpmInstall(fixturePath) {
   }
 }
 
-async function testPath(
-  isDev,
-  origin,
-  status,
-  path,
-  expectedText,
+export async function testPath(
+  isDev: boolean,
+  origin: string,
+  status: number,
+  path: string,
+  expectedText: string | Function | RegExp,
   expectedHeaders = {},
-  fetchOpts = {}
+  fetchOpts: FetchOptions = {}
 ) {
-  const opts = {
+  const opts: FetchOptions = {
     retries: isCI ? 5 : 0,
     ...fetchOpts,
+    // @ts-expect-error - this value is part of a hack to work around
+    // https://github.com/node-fetch/node-fetch/issues/417#issuecomment-587233352
     redirect: 'manual-dont-change',
     status,
   };
@@ -183,6 +205,7 @@ async function testPath(
   console.log(msg);
   expect(res.status).toBe(status);
   validateResponseHeaders(res);
+
   if (typeof expectedText === 'string') {
     const actualText = await res.text();
     expect(actualText.trim()).toBe(expectedText.trim());
@@ -194,6 +217,7 @@ async function testPath(
     expectedText.lastIndex = 0; // reset since we test twice
     expect(actualText).toMatch(expectedText);
   }
+
   if (expectedHeaders) {
     Object.entries(expectedHeaders).forEach(([key, expectedValue]) => {
       let actualValue = res.headers.get(key);
@@ -208,7 +232,11 @@ async function testPath(
   }
 }
 
-async function testFixture(directory, opts = {}, args = []) {
+export async function testFixture(
+  directory: string,
+  opts: Options<null> = {},
+  args: string[] = []
+) {
   await runNpmInstall(directory);
 
   const token = await fetchCachedToken();
@@ -238,7 +266,7 @@ async function testFixture(directory, opts = {}, args = []) {
       shell: true,
       stdio: 'pipe',
       ...opts,
-      env: { ...opts.env, __VERCEL_SKIP_DEV_CMD: 1 },
+      env: { ...opts.env, __VERCEL_SKIP_DEV_CMD: '1' },
     }
   );
 
@@ -246,6 +274,13 @@ async function testFixture(directory, opts = {}, args = []) {
   let stderr = '';
   const readyResolver = createResolver();
   const exitResolver = createResolver();
+
+  if (!dev.stdout) {
+    throw new Error('`vc dev` process missing "stdout".');
+  }
+  if (!dev.stderr) {
+    throw new Error('`vc dev` process missing "stderr".');
+  }
 
   dev.stdout.setEncoding('utf8');
   dev.stderr.setEncoding('utf8');
@@ -257,12 +292,12 @@ async function testFixture(directory, opts = {}, args = []) {
     stderr += data;
 
     if (stripAnsi(stderr).includes('Ready! Available at')) {
-      readyResolver.resolve();
+      readyResolver.resolve(null);
     }
   });
 
   let printedOutput = false;
-  let devTimer = null;
+  let devTimer: NodeJS.Timeout;
 
   dev.on('exit', code => {
     devTimer = setTimeout(async () => {
@@ -284,8 +319,8 @@ async function testFixture(directory, opts = {}, args = []) {
       printOutput(directory, stdout, stderr);
       printedOutput = true;
     }
-    exitResolver.resolve();
-    readyResolver.resolve();
+    exitResolver.resolve(null);
+    readyResolver.resolve(null);
   });
 
   dev.on('error', () => {
@@ -293,11 +328,11 @@ async function testFixture(directory, opts = {}, args = []) {
       printOutput(directory, stdout, stderr);
       printedOutput = true;
     }
-    exitResolver.resolve();
-    readyResolver.resolve();
+    exitResolver.resolve(null);
+    readyResolver.resolve(null);
   });
 
-  dev._kill = dev.kill;
+  // @ts-ignore
   dev.kill = async () => {
     // kill the entire process tree for the child as some tests will spawn
     // child processes that either become defunct or assigned a new parent
@@ -318,15 +353,15 @@ async function testFixture(directory, opts = {}, args = []) {
   };
 }
 
-function testFixtureStdio(
-  directory,
-  fn,
-  { skipDeploy, projectSettings, readyTimeout = 0 } = {}
+export function testFixtureStdio(
+  directory: string,
+  fn: Function,
+  { skipDeploy = false, projectSettings = {}, readyTimeout = 0 } = {}
 ) {
   return async () => {
     const cwd = fixtureAbsolute(directory);
     const token = await fetchCachedToken();
-    let deploymentUrl;
+    let deploymentUrl: string;
 
     // Deploy fixture and link project
     if (!skipDeploy) {
@@ -402,16 +437,15 @@ function testFixtureStdio(
           { cwd, stdio: 'pipe', reject: false }
         );
 
-        // eslint-disable-next-line no-console
-        console.log({
+        const errorDetails = JSON.stringify({
           exitCode: deployResult.exitCode,
           stdout: deployResult.stdout,
           stderr: deployResult.stderr,
         });
 
         // Expect the deploy succeeded with exit of 0;
-        expect(deployResult.exitCode).toBe(0);
-        deploymentUrl = new URL(deployResult.stdout).host;
+        expect(deployResult.exitCode, errorDetails).toBe(0);
+        deploymentUrl = new URL(deployResult.stdout.toString()).host;
       } finally {
         if (!hasGitignore) {
           await fs.remove(gitignore);
@@ -420,7 +454,7 @@ function testFixtureStdio(
     }
 
     // Start dev
-    let dev;
+    let dev: ExecaChildProcess<Buffer>;
 
     await runNpmInstall(cwd);
 
@@ -432,7 +466,7 @@ function testFixtureStdio(
     // By default, tests will wait 6 minutes for the dev server to be ready and
     // perform the tests, however a `readyTimeout` can be used to reduce the
     // wait time if the dev server is expected to fail to start or hang
-    let readyTimer = null;
+    let readyTimer: NodeJS.Timeout;
     if (readyTimeout > 0) {
       readyTimer = setTimeout(() => {
         readyResolver.reject(
@@ -451,14 +485,14 @@ function testFixtureStdio(
         } --debug`
       );
       const env = skipDeploy
-        ? { ...process.env, __VERCEL_SKIP_DEV_CMD: 1 }
+        ? { ...process.env, __VERCEL_SKIP_DEV_CMD: '1' }
         : process.env;
       dev = execa(
         binaryPath,
         [
           'dev',
           '-l',
-          port,
+          port.toString(),
           '-t',
           token,
           ...(process.env.VERCEL_TEAM_ID
@@ -471,6 +505,13 @@ function testFixtureStdio(
           env,
         }
       );
+
+      if (!dev.stdout) {
+        throw new Error('`vc dev` missing "stdout"');
+      }
+      if (!dev.stderr) {
+        throw new Error('`vc dev` missing "stderr"');
+      }
 
       dev.stdout.setEncoding('utf8');
       dev.stderr.setEncoding('utf8');
@@ -487,7 +528,7 @@ function testFixtureStdio(
 
         if (stripAnsi(data).includes('Ready! Available at')) {
           clearTimeout(readyTimer);
-          readyResolver.resolve();
+          readyResolver.resolve(null);
         }
 
         if (stderr.includes(`Requested port ${port} is already in use`)) {
@@ -508,7 +549,7 @@ function testFixtureStdio(
           printOutput(directory, stdout, stderr);
           printedOutput = true;
         }
-        exitResolver.resolve();
+        exitResolver.resolve(null);
       });
 
       dev.on('error', () => {
@@ -516,37 +557,41 @@ function testFixtureStdio(
           printOutput(directory, stdout, stderr);
           printedOutput = true;
         }
-        exitResolver.resolve();
+        exitResolver.resolve(null);
       });
 
       await readyResolver;
 
-      const helperTestPath = async (...args) => {
+      const helperTestPath = async (...args: any[]) => {
         if (!skipDeploy) {
+          // @ts-ignore
           await testPath(false, `https://${deploymentUrl}`, ...args);
         }
+        // @ts-ignore
         await testPath(true, `http://localhost:${port}`, ...args);
       };
       await fn(helperTestPath, port);
     } finally {
+      // @ts-ignore
       await nukeProcessTree(dev.pid);
       await exitResolver;
     }
   };
 }
 
-async function ps(parentPid, pids = {}) {
-  const cmd =
+async function ps(parentPid: number, pids: Record<string, Array<number>> = {}) {
+  const cmd: string[] =
     process.platform === 'darwin'
-      ? ['pgrep', '-P', parentPid]
-      : ['ps', '-o', 'pid', '--no-headers', '--ppid', parentPid];
+      ? ['pgrep', '-P', parentPid.toString()]
+      : ['ps', '-o', 'pid', '--no-headers', '--ppid', parentPid.toString()];
 
   try {
     const buf = execFileSync(cmd[0], cmd.slice(1), {
       encoding: 'utf-8',
     });
-    for (let pid of buf.match(/\d+/g)) {
-      pid = parseInt(pid);
+    const possiblePids = buf.match(/\d+/g) || [];
+    for (let rawPid of possiblePids) {
+      const pid = parseInt(rawPid);
       const recurse = Object.prototype.hasOwnProperty.call(pids, pid);
       pids[parentPid].push(pid);
       pids[pid] = [];
@@ -554,14 +599,19 @@ async function ps(parentPid, pids = {}) {
         await ps(pid, pids);
       }
     }
-  } catch (e) {
+  } catch (err) {
+    const error = err as Error;
     // eslint-disable-next-line no-console
-    console.log(`Failed to get processes: ${e.toString()}`);
+    console.log(`Failed to get processes: ${error.toString()}`);
   }
   return pids;
 }
 
-async function nukePID(pid, signal = 'SIGTERM', retries = 10) {
+async function nukePID(
+  pid: number,
+  signal: string = 'SIGTERM',
+  retries: number = 10
+) {
   if (retries === 0) {
     // eslint-disable-next-line no-console
     console.log(`pid ${pid} won't die, giving up`);
@@ -595,9 +645,11 @@ async function nukePID(pid, signal = 'SIGTERM', retries = 10) {
   await nukePID(pid, 'SIGKILL', retries - 1);
 }
 
-async function nukeProcessTree(pid, signal) {
+async function nukeProcessTree(pid: number, signal?: string) {
   if (process.platform === 'win32') {
-    spawnSync('taskkill', ['/pid', pid, '/T', '/F'], { stdio: 'inherit' });
+    spawnSync('taskkill', ['/pid', pid.toString(), '/T', '/F'], {
+      stdio: 'inherit',
+    });
     return;
   }
 
@@ -624,26 +676,16 @@ afterEach(async () => {
       try {
         await nukeProcessTree(proc.pid);
       } catch (err) {
+        const error = err as Error & { code?: string };
+
         // Was already killed
-        if (err.code !== 'ESRCH') {
+        if (error.code !== 'ESRCH') {
           // eslint-disable-next-line no-console
-          console.error('Failed to kill process', proc.pid, err);
+          console.error('Failed to kill process', proc.pid, error);
         }
       }
     })
   );
 });
 
-module.exports = {
-  sleep,
-  testPath,
-  testFixture,
-  testFixtureStdio,
-  exec,
-  formatOutput,
-  shouldSkip,
-  fixture,
-  fetch,
-  fetchWithRetry,
-  validateResponseHeaders,
-};
+export { fetch };