+++ packages/cli/test/mocks/project.ts
@@ -101,40 +101,36 @@ const systemEnvs = [
   },
 ];
 
+const latestProductionDeployment: Deployment = {
+  alias: ['foobar.com'],
+  aliasAssigned: 1571239348998,
+  buildingAt: 1571239348998,
+  createdAt: 1571239348998,
+  createdIn: 'sfo1',
+  creator: {
+    uid: 'K4amb7K9dAt5R2vBJWF32bmY',
+  },
+  forced: false,
+  id: 'dpl_89qyp1cskzkLrVicDaZoDbjyHuDJ',
+  meta: {},
+  plan: 'pro',
+  private: true,
+  readyState: 'READY',
+  target: 'production',
+  type: undefined,
+  url: 'a-project-name-rjtr4pz3f.vercel.app',
+};
 export const defaultProject: Project = {
   id: 'foo',
   name: 'cli',
   accountId: 'K4amb7K9dAt5R2vBJWF32bmY',
   createdAt: 1555413045188,
   updatedAt: 1555413045188,
-  latestDeployments: [
-    {
-      alias: ['foobar.com'],
-      aliasAssigned: 1571239348998,
-      buildingAt: 1571239348998,
-      createdAt: 1571239348998,
-      createdIn: 'sfo1',
-      creator: {
-        uid: 'K4amb7K9dAt5R2vBJWF32bmY',
-      },
-      forced: false,
-      id: 'dpl_89qyp1cskzkLrVicDaZoDbjyHuDJ',
-      meta: {},
-      plan: 'pro',
-      private: true,
-      readyState: 'READY',
-      target: 'production',
-      type: undefined,
-      url: 'a-project-name-rjtr4pz3f.vercel.app',
-    },
-  ],
+  latestDeployments: [latestProductionDeployment],
   lastAliasRequest: null,
-  alias: [
-    {
-      domain: 'foobar.com',
-      target: 'PRODUCTION' as const,
-    },
-  ],
+  targets: {
+    production: latestProductionDeployment,
+  },
 };
 
 /**
@@ -216,7 +212,7 @@ export function useProject(
     res.json(project);
   });
   client.scenario.get(
-    `/v1/env/pull/${project.id}/:target?/:gitBranch?`,
+    `/v2/env/pull/${project.id}/:target?/:gitBranch?`,
     (req, res) => {
       const target =
         typeof req.params.target === 'string'
@@ -351,7 +347,7 @@ export function useProject(
     }
     res.json(project);
   });
-  client.scenario.get(`/v4/projects`, (req, res) => {
+  client.scenario.get(`/v9/projects`, (req, res) => {
     res.json({
       projects: [project],
       pagination: null,
@@ -390,7 +386,7 @@ function exposeSystemEnvs(
 ) {
   const envs: Env = {};
 
-  if (autoExposeSystemEnvs) {
+  if (autoExposeSystemEnvs && target !== 'development') {
     envs['VERCEL'] = '1';
     envs['VERCEL_ENV'] = target || 'development';