+++ packages/next/src/utils.ts
@@ -192,8 +192,12 @@ function normalizePage(page: string): string {
   if (!page.startsWith('/')) {
     page = `/${page}`;
   }
-  // remove '/index' from the end
-  page = page.replace(/\/index$/, '/');
+
+  // Replace the `/index` with `/`
+  if (page === '/index') {
+    page = '/';
+  }
+
   return page;
 }
 
@@ -320,8 +324,8 @@ export async function getDynamicRoutes({
   bypassToken,
   isServerMode,
   dynamicMiddlewareRouteMap,
-  experimentalPPRRoutes,
   hasActionOutputSupport,
+  isAppPPREnabled,
 }: {
   entryPath: string;
   entryDirectory: string;
@@ -333,8 +337,8 @@ export async function getDynamicRoutes({
   bypassToken?: string;
   isServerMode?: boolean;
   dynamicMiddlewareRouteMap?: ReadonlyMap<string, RouteWithSrc>;
-  experimentalPPRRoutes: ReadonlySet<string>;
   hasActionOutputSupport: boolean;
+  isAppPPREnabled: boolean;
 }): Promise<RouteWithSrc[]> {
   if (routesManifest) {
     switch (routesManifest.version) {
@@ -408,7 +412,7 @@ export async function getDynamicRoutes({
             ];
           }
 
-          if (experimentalPPRRoutes.has(page)) {
+          if (isAppPPREnabled) {
             let dest = route.dest?.replace(/($|\?)/, '.prefetch.rsc$1');
 
             if (page === '/' || page === '/index') {
@@ -1439,6 +1443,13 @@ async function getSourceFilePathFromPage({
     }
   }
 
+  // if we got here, and didn't find a source not-found file, then it was the one injected
+  // by Next.js. There's no need to warn or return a source file in this case, as it won't have
+  // any configuration applied to it.
+  if (page === '/_not-found/page') {
+    return '';
+  }
+
   console.log(
     `WARNING: Unable to find source file for page ${page} with extensions: ${extensionsToTry.join(
       ', '
@@ -1498,8 +1509,8 @@ export type LambdaGroup = {
   isAppRouter?: boolean;
   isAppRouteHandler?: boolean;
   isStreaming?: boolean;
-  isPrerenders?: boolean;
-  isExperimentalPPR?: boolean;
+  readonly isPrerenders: boolean;
+  readonly isExperimentalPPR: boolean;
   isActionLambda?: boolean;
   isPages?: boolean;
   isApiLambda: boolean;
@@ -1565,15 +1576,13 @@ export async function getPageLambdaGroups({
     }
 
     if (config && config.functions) {
-      // `pages` are normalized without route groups (e.g., /app/(group)/page.js).
-      // we keep track of that mapping in `inversedAppPathManifest`
-      // `getSourceFilePathFromPage` needs to use the path from source to properly match the config
-      const pageFromManifest = inversedAppPathManifest?.[routeName];
       const sourceFile = await getSourceFilePathFromPage({
         workPath: entryPath,
-        // since this function is used by both `pages` and `app`, the manifest might not be provided
-        // so fallback to normal behavior of just checking the `page`.
-        page: pageFromManifest ?? page,
+        page: normalizeSourceFilePageFromManifest(
+          routeName,
+          page,
+          inversedAppPathManifest
+        ),
         pageExtensions,
       });
 
@@ -1655,6 +1664,43 @@ export async function getPageLambdaGroups({
   }
 
   return groups;
+}
+
+// `pages` are normalized without route groups (e.g., /app/(group)/page.js).
+// we keep track of that mapping in `inversedAppPathManifest`
+// `getSourceFilePathFromPage` needs to use the path from source to properly match the config
+function normalizeSourceFilePageFromManifest(
+  routeName: string,
+  page: string,
+  inversedAppPathManifest?: Record<string, string>
+) {
+  const pageFromManifest = inversedAppPathManifest?.[routeName];
+  if (!pageFromManifest) {
+    // since this function is used by both `pages` and `app`, the manifest might not be provided
+    // so fallback to normal behavior of just checking the `page`.
+    return page;
+  }
+
+  const metadataConventions = [
+    '/favicon.',
+    '/icon.',
+    '/apple-icon.',
+    '/opengraph-image.',
+    '/twitter-image.',
+    '/sitemap.',
+    '/robots.',
+  ];
+
+  // these special metadata files for will not contain `/route` or `/page` suffix, so return the routeName as-is.
+  const isSpecialFile = metadataConventions.some(convention =>
+    routeName.startsWith(convention)
+  );
+
+  if (isSpecialFile) {
+    return routeName;
+  }
+
+  return pageFromManifest;
 }
 
 export const outputFunctionFileSizeInfo = (
@@ -1906,6 +1952,8 @@ type OnPrerenderRouteArgs = {
   routesManifest?: RoutesManifest;
   isCorrectNotFoundRoutes?: boolean;
   isEmptyAllowQueryForPrendered?: boolean;
+  isAppPPREnabled: boolean;
+  hasActionOutputSupport?: boolean;
 };
 let prerenderGroup = 1;
 
@@ -1942,6 +1990,8 @@ export const onPrerenderRoute =
       routesManifest,
       isCorrectNotFoundRoutes,
       isEmptyAllowQueryForPrendered,
+      isAppPPREnabled,
+      hasActionOutputSupport,
     } = prerenderRouteArgs;
 
     if (isBlocking && isFallback) {
@@ -2062,14 +2112,14 @@ export const onPrerenderRoute =
 
     // If enabled, try to get the postponed route information from the file
     // system and use it to assemble the prerender.
-    let prerender: string | undefined;
+    let postponedPrerender: string | undefined;
     if (experimentalPPR && appDir) {
       const htmlPath = path.join(appDir, `${routeFileNoExt}.html`);
       const metaPath = path.join(appDir, `${routeFileNoExt}.meta`);
       if (fs.existsSync(htmlPath) && fs.existsSync(metaPath)) {
         const meta = JSON.parse(await fs.readFile(metaPath, 'utf8'));
         if ('postponed' in meta && typeof meta.postponed === 'string') {
-          prerender = meta.postponed;
+          postponedPrerender = meta.postponed;
 
           // Assign the headers Content-Type header to the prerendered type.
           initialHeaders ??= {};
@@ -2079,7 +2129,7 @@ export const onPrerenderRoute =
 
           // Read the HTML file and append it to the prerendered content.
           const html = await fs.readFileSync(htmlPath, 'utf8');
-          prerender += html;
+          postponedPrerender += html;
         }
       }
 
@@ -2096,14 +2146,14 @@ export const onPrerenderRoute =
       }
     }
 
-    if (prerender) {
+    if (postponedPrerender) {
       const contentType = initialHeaders?.['content-type'];
       if (!contentType) {
         throw new Error("Invariant: contentType can't be undefined");
       }
 
       // Assemble the prerendered file.
-      htmlFsRef = new FileBlob({ contentType, data: prerender });
+      htmlFsRef = new FileBlob({ contentType, data: postponedPrerender });
     } else if (
       appDir &&
       !dataRoute &&
@@ -2167,7 +2217,14 @@ export const onPrerenderRoute =
                     ? addLocaleOrDefault('/404.html', routesManifest, locale)
                     : '/404.html'
                   : isAppPathRoute
-                  ? prefetchDataRoute || dataRoute
+                  ? // When experimental PPR is enabled, we expect that the data
+                    // that should be served as a part of the prerender should
+                    // be from the prefetch data route. If this isn't enabled
+                    // for ppr, the only way to get the data is from the data
+                    // route.
+                    experimentalPPR
+                    ? prefetchDataRoute
+                    : dataRoute
                   : routeFileNoExt + '.json'
               }`
             ),
@@ -2175,22 +2232,6 @@ export const onPrerenderRoute =
 
     if (isOmittedOrNotFound) {
       initialStatus = 404;
-    }
-
-    /**
-     * If the route key had an `/index` suffix added, we need to note it so we
-     * can remove it from the output path later accurately.
-     */
-    let addedIndexSuffix = false;
-
-    if (isAppPathRoute) {
-      // for literal index routes we need to append an additional /index
-      // due to the proxy's normalizing for /index routes
-      if (routeKey !== '/index' && routeKey.endsWith('/index')) {
-        routeKey = `${routeKey}/index`;
-        routeFileNoExt = routeKey;
-        addedIndexSuffix = true;
-      }
     }
 
     let outputPathPage = path.posix.join(entryDirectory, routeFileNoExt);
@@ -2229,7 +2270,7 @@ export const onPrerenderRoute =
 
     let outputPathPrefetchData: null | string = null;
     if (prefetchDataRoute) {
-      if (!experimentalPPR) {
+      if (!isAppPPREnabled) {
         throw new Error(
           "Invariant: prefetchDataRoute can't be set without PPR"
         );
@@ -2239,10 +2280,6 @@ export const onPrerenderRoute =
     } else if (experimentalPPR) {
       throw new Error('Invariant: expected to find prefetch data route PPR');
     }
-
-    // When the prefetch data path is available, use it for the prerender,
-    // otherwise use the data path.
-    const outputPrerenderPathData = outputPathPrefetchData || outputPathData;
 
     if (isSharedLambdas) {
       const outputSrcPathPage = normalizeIndexOutput(
@@ -2296,8 +2333,14 @@ export const onPrerenderRoute =
         htmlFsRef.contentType = htmlContentType;
         prerenders[outputPathPage] = htmlFsRef;
 
-        if (outputPrerenderPathData) {
-          prerenders[outputPrerenderPathData] = jsonFsRef;
+        if (outputPathPrefetchData) {
+          prerenders[outputPathPrefetchData] = jsonFsRef;
+        }
+
+        // If experimental ppr is not enabled for this route, then add the data
+        // route as a target for the prerender as well.
+        if (outputPathData && !experimentalPPR) {
+          prerenders[outputPathData] = jsonFsRef;
         }
       }
     }
@@ -2386,7 +2429,7 @@ export const onPrerenderRoute =
         // static route first, then try the srcRoute if it doesn't exist. If we
         // can't find it at all, this constitutes an error.
         experimentalStreamingLambdaPath = experimentalStreamingLambdaPaths.get(
-          pathnameToOutputName(entryDirectory, routeKey, addedIndexSuffix)
+          pathnameToOutputName(entryDirectory, routeKey)
         );
         if (!experimentalStreamingLambdaPath && srcRoute) {
           experimentalStreamingLambdaPath =
@@ -2433,21 +2476,27 @@ export const onPrerenderRoute =
           : {}),
       });
 
-      if (outputPrerenderPathData) {
-        let normalizedPathData = outputPrerenderPathData;
+      if (hasActionOutputSupport) {
+        const actionOutputKey = `${path.join('./', srcRoute || '')}.action`;
+        if (srcRoute !== routeKey && lambdas[actionOutputKey]) {
+          lambdas[`${routeKey}.action`] = lambdas[actionOutputKey];
+        }
+      }
 
+      const normalizePathData = (pathData: string) => {
         if (
           (srcRoute === '/' || srcRoute == '/index') &&
-          outputPrerenderPathData.endsWith(RSC_PREFETCH_SUFFIX)
+          pathData.endsWith(RSC_PREFETCH_SUFFIX)
         ) {
-          delete lambdas[normalizedPathData];
-          normalizedPathData = normalizedPathData.replace(
-            /([^/]+\.prefetch\.rsc)$/,
-            '__$1'
-          );
+          delete lambdas[pathData];
+          return pathData.replace(/([^/]+\.prefetch\.rsc)$/, '__$1');
         }
 
-        prerenders[normalizedPathData] = new Prerender({
+        return pathData;
+      };
+
+      if (outputPathData || outputPathPrefetchData) {
+        const prerender = new Prerender({
           expiration: initialRevalidate,
           lambda,
           allowQuery,
@@ -2468,21 +2517,30 @@ export const onPrerenderRoute =
                   ...initialHeaders,
                   'content-type': rscContentTypeHeader,
                   vary: rscVaryHeader,
-                  // If it contains a pre-render, then it was postponed.
-                  ...(prerender && rscDidPostponeHeader
+                  ...(postponedPrerender && rscDidPostponeHeader
                     ? { [rscDidPostponeHeader]: '1' }
                     : {}),
                 },
               }
             : {}),
         });
+
+        if (outputPathPrefetchData) {
+          prerenders[normalizePathData(outputPathPrefetchData)] = prerender;
+        }
+
+        // If experimental ppr is not enabled for this route, then add the data
+        // route as a target for the prerender as well.
+        if (outputPathData && !experimentalPPR) {
+          prerenders[normalizePathData(outputPathData)] = prerender;
+        }
       }
 
       // we need to ensure all prerenders have a matching .rsc output
       // otherwise routing could fall through unexpectedly for the
       // fallback: false case as it doesn't have a dynamic route
       // to catch the `.rsc` request for app -> pages routing
-      if (outputPrerenderPathData?.endsWith('.json') && appDir) {
+      if (outputPathData?.endsWith('.json') && appDir) {
         const dummyOutput = new FileBlob({
           data: '{}',
           contentType: 'application/json',
@@ -2647,18 +2705,9 @@ export function getNextServerPath(nextVersion: string) {
     : 'next/dist/next-server/server';
 }
 
-function pathnameToOutputName(
-  entryDirectory: string,
-  pathname: string,
-  addedIndexSuffix = false
-) {
+function pathnameToOutputName(entryDirectory: string, pathname: string) {
   if (pathname === '/') {
     pathname = '/index';
-  }
-  // If the `/index` was added for a route that ended in `/index` we need to
-  // strip the second one off before joining it with the entryDirectory.
-  else if (addedIndexSuffix) {
-    pathname = pathname.replace(/\/index$/, '');
   }
 
   return path.posix.join(entryDirectory, pathname);