+++ packages/fs-detectors/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @vercel/fs-detectors
 
+## 5.2.5
+
+### Patch Changes
+
+- Updated dependencies [[`1484df7aa`](https://github.com/vercel/vercel/commit/1484df7aa55262646544449d26762a7a92d89a2e)]:
+  - @vercel/frameworks@3.0.3
+
+## 5.2.4
+
+### Patch Changes
+
+- Add support for detecting Turborepo 2 ([#11680](https://github.com/vercel/vercel/pull/11680))
+
 ## 5.2.3
 
 ### Patch Changes